requirements/base.txt
asgiref==3.7.2
astroid==3.0.2
colorama==0.4.6
dill==0.3.8
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.5
django-safedelete==1.3.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-writable-nested==0.7.0
drf-yasg==1.21.7
inflection==0.5.1
isort==5.13.2
mccabe==0.7.0
packaging==23.2
pillow==10.2.0
platformdirs==4.2.0
PyJWT==2.8.0
pylint==3.0.3
pytz==2023.4
PyYAML==6.0.1
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
